// src/app/models/jobresponse.ts
export interface jobsResponses {
  data?: (DataEntity)[] | null;
  links: Links;
  total_count: number;
}
export interface DataEntity {
  id: string;
  type: string;
  recipient_sender_device_signature_file_url?: null;
  signature_file_url?: null;
  photo_1_file_url?: string | null;
  photo_2_file_url?: string | null;
  photo_3_file_url?: null;
  photo_4_file_url?: null;
  photo_5_file_url?: null;
  photo_6_file_url?: null;
  photo_7_file_url?: null;
  photo_8_file_url?: null;
  photo_9_file_url?: null;
  photo_10_file_url?: null;
  primary_job_status: string;
  open_to_marketplace?: boolean | null;
  marketplace_offer?: null;
  do_number: string;
  attempt: number;
  date: string;
  start_date: string;
  job_age: number;
  job_release_time?: null;
  job_time?: null;
  time_window?: null;
  job_received_date?: null;
  tracking_number?: string | null;
  order_number?: string | null;
  job_type?: null;
  job_sequence?: null;
  job_fee?: null;
  address_lat?: null;
  address_lng?: null;
  address: string;
  company_name?: string | null;
  address_1?: string | null;
  address_2?: string | null;
  address_3?: null;
  postal_code?: null;
  city?: string | null;
  state?: string | null;
  country?: string | null;
  billing_address?: null;
  deliver_to_collect_from?: string | null;
  last_name?: null;
  phone_number?: string | null;
  sender_phone_number?: string | null;
  fax_number?: null;
  instructions?: string | null;
  assign_to?: string | null;
  notify_email?: string | null;
  webhook_url?: null;
  zone?: null;
  customer?: string | null;
  account_number?: string | null;
  job_owner?: null;
  invoice_number?: null;
  invoice_amount?: null;
  payment_mode?: null;
  payment_amount?: number | null;
  group_id?: null;
  group_name?: null;
  group_code?: null;
  source?: null;
  weight?: number | null;
  parcel_width?: null;
  parcel_length?: null;
  parcel_height?: null;
  cubic_meters?: null;
  boxes?: null;
  cartons?: null;
  pieces?: number | null;
  envelopes?: null;
  pallets?: null;
  bins?: null;
  trays?: null;
  bundles?: null;
  rolls?: null;
  number_of_shipping_labels?: null;
  attachment_url?: null;
  detrack_number: string;
  status: string;
  tracking_status: string;
  reason?: string | null;
  received_by_sent_by?: string | null;
  note?: string | null;
  carrier?: null;
  pod_time?: string | null;
  pod_lat: number | string;
  pod_lng: number | string;
  pod_address: string;
  address_tracked_at?: string | null;
  arrived_lat?: null;
  arrived_lng?: null;
  arrived_address?: null;
  arrived_at?: null;
  texted_at?: null;
  called_at?: null;
  serial_number?: string | null;
  recipient_sender_device_pod_at?: null;
  recipient_sender_device_signed_by?: null;
  signed_at?: null;
  photo_1_at?: string | null;
  photo_2_at?: string | null;
  photo_3_at?: null;
  photo_4_at?: null;
  photo_5_at?: null;
  photo_6_at?: null;
  photo_7_at?: null;
  photo_8_at?: null;
  photo_9_at?: null;
  photo_10_at?: null;
  actual_weight?: null;
  temperature?: number | null;
  hold_time?: null;
  payment_collected?: null;
  auto_reschedule?: null;
  actual_crates?: null;
  actual_pallets?: null;
  actual_utilization?: null;
  goods_service_rating?: null;
  driver_rating?: null;
  customer_feedback?: null;
  eta_time?: null;
  live_eta?: null;
  depot?: null;
  depot_contact?: null;
  department?: null;
  sales_person?: null;
  identification_number?: null;
  bank_prefix?: null;
  run_number?: null;
  head_to_delivery_at?: string | null;
  pod_at?: string | null;
  job_price?: null;
  insurance_price?: null;
  insurance_coverage: boolean;
  total_price?: null;
  payer_type?: null;
  remarks?: null;
  items_count: number;
  service_type?: null;
  warehouse_address?: null;
  destination_time_window?: null;
  door?: null;
  time_zone?: null;
  created_at: string;
  on_demand: boolean;
  vehicle_type?: null;
  priority?: null;
  service_time?: null;
  contractor_group_name: string;
  connect_token?: null;
  connect_host?: null;
  connect_id?: null;
  job_order?: null;
  tracking_status_code: string;
  driver_mobile_number?: string | null;
  reattempted?: null;
  cubic_meter?: null;
  tracking_link: string;
  contactless_signature_link: string;
  available_statuses: AvailableStatuses;
  items?: (null)[] | null;
  milestones?: (MilestonesEntity)[] | null;
}
export interface AvailableStatuses {
  info_recv: string;
  dispatched: string;
  completed: string;
  completed_partial: string;
  failed: string;
  on_hold: string;
  return: string;
}
export interface MilestonesEntity {
  status: string;
  assign_to?: string | null;
  reason?: string | null;
  pod_at: string;
  created_at: string;
  user_name?: string | null;
}
export interface Links {
  first: string;
  last: string;
  next: string;
  prev?: null;
}